+++ egui/src/ui.rs
@@ -814,13 +814,35 @@ impl Ui {
     /// Shows a button with the given color.
     /// If the user clicks the button, a full color picker is shown.
     pub fn color_edit_button_srgba(&mut self, srgba: &mut Color32) -> Response {
-        widgets::color_picker::color_edit_button_srgba(self, srgba)
+        color_picker::color_edit_button_srgba(self, srgba, color_picker::Alpha::BlendOrAdditive)
     }
 
     /// Shows a button with the given color.
     /// If the user clicks the button, a full color picker is shown.
     pub fn color_edit_button_hsva(&mut self, hsva: &mut Hsva) -> Response {
-        widgets::color_picker::color_edit_button_hsva(self, hsva)
+        color_picker::color_edit_button_hsva(self, hsva, color_picker::Alpha::BlendOrAdditive)
+    }
+
+    /// Shows a button with the given color.
+    /// If the user clicks the button, a full color picker is shown.
+    /// The given color is in `sRGB` space.
+    pub fn color_edit_button_srgb(&mut self, srgb: &mut [u8; 3]) -> Response {
+        let mut hsva = Hsva::from_srgb(*srgb);
+        let response =
+            color_picker::color_edit_button_hsva(self, &mut hsva, color_picker::Alpha::Opaque);
+        *srgb = hsva.to_srgb();
+        response
+    }
+
+    /// Shows a button with the given color.
+    /// If the user clicks the button, a full color picker is shown.
+    /// The given color is in linear RGB space.
+    pub fn color_edit_button_rgb(&mut self, rgb: &mut [f32; 3]) -> Response {
+        let mut hsva = Hsva::from_rgb(*rgb);
+        let response =
+            color_picker::color_edit_button_hsva(self, &mut hsva, color_picker::Alpha::Opaque);
+        *rgb = hsva.to_rgb();
+        response
     }
 
     /// Shows a button with the given color.
@@ -839,7 +861,8 @@ impl Ui {
     /// If unsure, what "premultiplied alpha" is, then this is probably the function you want to use.
     pub fn color_edit_button_srgba_unmultiplied(&mut self, srgba: &mut [u8; 4]) -> Response {
         let mut hsva = Hsva::from_srgba_unmultiplied(*srgba);
-        let response = self.color_edit_button_hsva(&mut hsva);
+        let response =
+            color_picker::color_edit_button_hsva(self, &mut hsva, color_picker::Alpha::OnlyBlend);
         *srgba = hsva.to_srgba_unmultiplied();
         response
     }
@@ -849,7 +872,11 @@ impl Ui {
     /// The given color is in linear RGBA space with premultiplied alpha
     pub fn color_edit_button_rgba_premultiplied(&mut self, rgba: &mut [f32; 4]) -> Response {
         let mut hsva = Hsva::from_rgba_premultiplied(*rgba);
-        let response = self.color_edit_button_hsva(&mut hsva);
+        let response = color_picker::color_edit_button_hsva(
+            self,
+            &mut hsva,
+            color_picker::Alpha::BlendOrAdditive,
+        );
         *rgba = hsva.to_rgba_premultiplied();
         response
     }
@@ -860,7 +887,8 @@ impl Ui {
     /// If unsure, what "premultiplied alpha" is, then this is probably the function you want to use.
     pub fn color_edit_button_rgba_unmultiplied(&mut self, rgba: &mut [f32; 4]) -> Response {
         let mut hsva = Hsva::from_rgba_unmultiplied(*rgba);
-        let response = self.color_edit_button_hsva(&mut hsva);
+        let response =
+            color_picker::color_edit_button_hsva(self, &mut hsva, color_picker::Alpha::OnlyBlend);
         *rgba = hsva.to_rgba_unmultiplied();
         response
     }